+++ client/components/Search.jsx
@@ -1,0 +1,20 @@
+import React from "react";
+import { Link } from "react-router-dom";
+
+// mui components
+import { Button } from "@mui/material";
+
+const Search = (props) => {
+  
+  return (
+    <div>
+      <Link to="/search/feed">
+        <Button variant="contained" onClick={()=> props.searchRecipes()}>
+            Find recipes
+        </Button>
+      </Link>
+    </div>
+  );
+}
+
+export default Search;
